requirements.txt
Django>=3.0,<4.0
uWSGI>=2.0.18,<2.1
djangorestframework
djongo==1.3.3
pytz==2024.1
django-cassandra-engine==1.6.1
celery
celery-beat